Extract conversion of positional to named placeholder to an utility class
